requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
gunicorn==21.2.0